fix fee tests and add xcm v4 tests
