+++ test/suites/dev/moonbase/test-xcm-v4/test-mock-dmp-error-and-appendix-4.ts
@@ -15,8 +15,8 @@ const RELAY_TOKEN = 1_000_000_000_000n;
 const palletId = "0x6D6f646c617373746d6E67720000000000000000";
 
 describeSuite({
-  id: "D013906",
-  title: "Mock XCM V3 - downward transfer claim trapped assets",
+  id: "D013904",
+  title: "Mock XCM V3 - downward transfer with always triggered appendix",
   foundationMethods: "dev",
   testCases: ({ context, it, log }) => {
     let assetId: string;
@@ -30,52 +30,11 @@ describeSuite({
       );
       assetId = registeredAssetId;
       expect(registeredAsset.owner.toHex()).to.eq(palletId.toLowerCase());
-
-      // BuyExecution does not charge for fees because we registered it for not doing so
-      // But since there is no error, and the deposit is on the error handler, the assets
-      // will be trapped.
-      // Goal is to trapp assets, so that later can be claimed
-      // Since we only BuyExecution, but we do not do anything with the assets after that,
-      // they are trapped
-      const xcmMessage = new XcmFragment({
-        assets: [
-          {
-            multilocation: {
-              parents: 1,
-              interior: {
-                Here: null,
-              },
-            },
-            fungible: 10n * RELAY_TOKEN,
-          },
-        ],
-      })
-        .reserve_asset_deposited()
-        .buy_execution()
-        .as_v2();
-
-      const receivedMessage: XcmVersionedXcm = context
-        .polkadotJs()
-        .createType("XcmVersionedXcm", xcmMessage);
-
-      const totalMessage = [...receivedMessage.toU8a()];
-      // Send RPC call to inject XCM message
-      await customDevRpcRequest("xcm_injectDownwardMessage", [totalMessage]);
-
-      // Create a block in which the XCM will be executed
-      await context.createBlock();
-
-      // Make sure ALITH did not reveive anything
-      const alith_dot_balance = await context
-        .polkadotJs()
-        .query.assets.account(assetId, alith.address);
-
-      expect(alith_dot_balance.isNone).to.be.true;
     });
 
     it({
       id: "T01",
-      title: "Should make sure that Alith receives claimed assets",
+      title: "Should make sure Alith receives 10 dot with appendix and error",
       test: async function () {
         const xcmMessage = new XcmFragment({
           assets: [
@@ -91,21 +50,22 @@ describeSuite({
           ],
           beneficiary: alith.address,
         })
-          // Claim assets that were previously trapped
-          // assets: the assets that were trapped
-          // ticket: the version of the assets (xcm version)
-          .claim_asset()
+          .reserve_asset_deposited()
           .buy_execution()
-          // Deposit assets, this time correctly, on Alith
-          .deposit_asset()
-          .as_v2();
+          // BuyExecution does not charge for fees because we registered it for not doing so
+          // As a consequence the trapped assets will be entirely credited
+          // The goal is to show appendix runs even if there is an error
+          .with(function () {
+            return this.set_appendix_with([this.deposit_asset_v3]);
+          })
+          .trap()
+          .as_v4();
 
         const receivedMessage: XcmVersionedXcm = context
           .polkadotJs()
           .createType("XcmVersionedXcm", xcmMessage);
 
         const totalMessage = [...receivedMessage.toU8a()];
-
         // Send RPC call to inject XCM message
         await customDevRpcRequest("xcm_injectDownwardMessage", [totalMessage]);
